pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-reconciliation"
version = "0.1.0"
description = "Modulares Tool für lineare stationäre Datenrekonziliation von Prozessdaten"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pandas",
    "openpyxl",
    "matplotlib",
    "scipy",
]

[project.optional-dependencies]
cuda = ["cupy-cuda12x"]
dev  = ["pytest", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
